requirements.txt
fastapi
uvicorn
pyyaml
qiskit
pydantic
numpy
